+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/cds-intent/panel-intent-header/panel-intent-header.component.ts
@@ -14,6 +14,7 @@ export class PanelIntentHeaderComponent implements OnInit, OnChanges {
   @ViewChild('myInput', { static: true }) myInput!: ElementRef<HTMLInputElement>;
 
   @Input() intent: Intent;
+  @Input() intentColor: string;
   @Output() saveIntent = new EventEmitter();
 
   listOfIntents: Intent[];