+++ src/app/models/intent-model.ts
@@ -1,5 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
 import { Action } from './action-model';
+import { INTENT_COLORS } from 'src/app/chatbot-design-studio/utils';
 
 export class Intent {
     webhook_enabled?: boolean;
@@ -34,6 +35,7 @@ export class IntentAttributes {
     position?: any;
     nextBlockAction?: any;
     connectors?: any;
+    color?: any;
     constructor() {
         this.position = {x:0, y:0};
         this.nextBlockAction = {
@@ -42,6 +44,7 @@ export class IntentAttributes {
             intentName: ""
         };
         this.connectors = {}
+        this.color = INTENT_COLORS.COLOR1
     }
 }