+++ CHANGELOG.md
@@ -7,7 +7,23 @@
 ### **Copyrigth**: 
 *Tiledesk SRL*
 
-# 1.26.0
+
+# 1.27.0-rc.2
+- **bug-fixed**: change connector default color and alpha  
+
+# 1.27.0-rc.1
+- **added**: added panel to change intent and connector colors
+
+# 1.26.0-rc.14
+- **added**: management for free expired plan
+
+# 1.26.0-rc.13
+- **bug-fixed**: copy and paste intent/action from different chatbots; 
+- **bug-fixed**: added vertical scroll to bot details page;
+
+# 1.26.0-rc.12
+- **added**: save the stage settings to local storage and load it at startup
+- **bug-fixed**: set the position to the top of the intent when the element height is greater than the stage size; 
 
 # 1.26.0-rc.11
 - **bug-fixed**: hide connector on mouseover only if alphaConnectors is 0